RDKB-36610: Debug
Reason for change: for onboarding
Test Procedure: Debug
Priority: P0
Risks: Low

+++ source/stats/wifi_monitor.c
@@ -1390,8 +1390,8 @@ int get_nop_started_channels(wifi_mon_stats_config_t *data)
         for (int i = 0; i < channels_num; i++) {
             bool found = false;
 
-            for (int j = 0; j < g_monitor_module.nop_channels_num; j++) {
-                if (g_monitor_module.nop_started_channels[j] == channel_list[i]) {
+            for (unsigned int j = 0; j < g_monitor_module.nop_channels_num; j++) {
+                if ((int)g_monitor_module.nop_started_channels[j] == channel_list[i]) {
                     found = true;
                     break;
                 }
@@ -1410,7 +1410,7 @@ int get_nop_started_channels(wifi_mon_stats_config_t *data)
                 }
 
                 g_monitor_module.nop_started_channels = new_ptr;
-                g_monitor_module.nop_started_channels[g_monitor_module.nop_channels_num] = channel_list[i];
+                (int)g_monitor_module.nop_started_channels[g_monitor_module.nop_channels_num] = channel_list[i];
                 g_monitor_module.nop_channels_num++;
 
                 wifi_util_dbg_print(WIFI_CTRL, "%s:%d Added NOP Channel: %u\n", __func__, __LINE__, channel_list[i]);
@@ -1419,10 +1419,10 @@ int get_nop_started_channels(wifi_mon_stats_config_t *data)
     } else {
         // Remove channels that have completed NOP
         for (int i = 0; i < channels_num; i++) {
-            for (int j = 0; j < g_monitor_module.nop_channels_num; j++) {
-                if (g_monitor_module.nop_started_channels[j] == channel_list[i]) {
+            for (unsigned int j = 0; j < g_monitor_module.nop_channels_num; j++) {
+                if ((int)g_monitor_module.nop_started_channels[j] == channel_list[i]) {
                     // Shift remaining entries
-                    for (int k = j; k < g_monitor_module.nop_channels_num - 1; k++) {
+                    for (unsigned int k = j; k < g_monitor_module.nop_channels_num - 1; k++) {
                         g_monitor_module.nop_started_channels[k] = g_monitor_module.nop_started_channels[k + 1];
                     }
 
@@ -1445,7 +1445,7 @@ int get_nop_started_channels(wifi_mon_stats_config_t *data)
     }
 
     wifi_util_dbg_print(WIFI_CTRL, "%s:%d Final NOP Channel Count: %d\n", __func__, __LINE__, g_monitor_module.nop_channels_num);
-    for (int j = 0; j < g_monitor_module.nop_channels_num; j++) {
+    for (unsigned int j = 0; j < g_monitor_module.nop_channels_num; j++) {
         wifi_util_dbg_print(WIFI_CTRL, "%s:%d Channel[%d] = %u\n", __func__, __LINE__, j, g_monitor_module.nop_started_channels[j]);
     }
